reference/requirements.txt
opencv-python
pyzbar
matplotlib
numpy
Pillow
qrcode[pil]
configparser
mysql-connector-python 